requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
firebase-admin==6.2.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0
safety==2.3.5
bandit==1.7.5